+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added Ink display UI.
 - Added Nessie SC
 - Added a pet level up alert
+- Added a reindrake alert
 
 ### Changes:
 - Enhanced `/rfureset` command to reset all session data, including fishing XP, ink tracking, and sea creature stats.